Set up email message with recipient, subject, and body
Added a new `MailboxAddress` for "waisa" with the email "alnymybdalwas1@gmail.com" to the `To` field. Set the email subject to "Hi waisa". Created a `BodyBuilder` to add an attachment "test.png", set the text body to "Hello Waisa", and the HTML body to a greeting and well wishes from Abdulwais Alnuaimi. Assigned the constructed body to `message.Body`.